python/pyproject.toml
[tool.poetry]
name = "algorise-geo-job-search-main"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pylint = "^3.1.0"
pytest = "^8.1.1"
pytest-cov = "^4.1.0"
requests = "^2.26.0"
beautifulsoup4 = "^4.10.0"
mysql-connector-python = "^8.0.28"
lxml = "^4.7.3"
selenium = "^4.1.0"
webdriver-manager = "^3.4.2"
pytest-mock = "^3.12.0"
pluggy = "^1.4.0"
googlemaps = "^4.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
